Also refresh automatically on the frontend
