+++ dashboard/client/components/GithubTeamIssueTable.module.scss
@@ -1,0 +1,33 @@
+.container {
+  padding: 1rem;
+
+  table {
+    border-spacing: 0;
+    border-left: 1px solid black;
+    border-top: 1px solid black;
+
+    tr {
+      :last-child {
+        td {
+          border-bottom: 0;
+        }
+      }
+    }
+
+    th,
+    td {
+      margin: 0;
+      padding: 0.5rem;
+      border-bottom: 1px solid black;
+      border-right: 1px solid black;
+
+      :last-child {
+        border-right: 0;
+      }
+    }
+
+    .stats {
+      text-align: center;
+    }
+  }
+}
